refactor: extract shared prompt functions from init wizard
Move prompt_single_option and prompt_feature_options into a shared
features/prompts.rs module so both cella init wizard and the upcoming
cella features edit can reuse them.
