+++ crates/cella-cli/src/commands/features/mod.rs
@@ -1,0 +1,4 @@
+//! `cella features` subcommands: edit, list, and update features in
+//! devcontainer configurations.
+
+pub mod prompts;
